requirements-dev.txt
pytest==3.2.3
pytest-cov==2.5.1
pytest-django==3.1.2
django-dynamic-fixture==2.0.0
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
